pyproject.toml
[tool.poetry]
name = "tartan-weaver"
version = "0.1.0"
description = "Creating Tartan images"
authors = ["drummonds <hum3@drummond.info>"]

[tool.poetry.dependencies]
python = "^3.8"
svgwrite = "^1.4.2"
Pillow = "^9.1.1"
black = {extras = ["jupyter"], version = "^22.3.0"}
shiv = "^1.0.3"
argparse = "^1.4.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
jupyterlab = "^3.4.3"
pytest = "^7.1.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
